# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "description": "Public self-signup. Creates a read-only `public` user with a bcrypt-hashed password (cost 12) and immediately issues access + refresh tokens (signup logs you in). Username must be unique. Rate limited to 5/min/IP.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created; tokens issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Username is already taken."
          }
        },
        "summary": "Register a public account (email + password)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Password login for self-registered public users. Admin accounts must use `POST /auth/self` (password + TOTP). Account locks for 15 min after 5 failures. Rate limited to 10/min/IP.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials."
          }
        },
        "summary": "Login with username + password (PUBLIC)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/tang-rat": {
      "post": {
        "description": "Verifies the mToken with the Tang Rat provider, finds or creates the user, and issues access + refresh tokens. Sets the refresh token as an httpOnly cookie.",
        "operationId": "AuthController_tangRat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TangRatLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid mToken or inactive user."
          }
        },
        "summary": "Login via ทางรัฐ mToken (PUBLIC / INSPECTOR / SUPERVISOR)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/self": {
      "post": {
        "description": "Admin self-login. On first login (mustChangePassword) returns a `tempToken` and `requiresPasswordChange: true` instead of tokens. Rate limited to 10/min/IP. Account locks for 15 min after 5 failures.",
        "operationId": "AuthController_self",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens, or a password-change challenge on first login.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or TOTP."
          }
        },
        "summary": "Login with username + password + TOTP (ADMIN only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Issues a new access + refresh token pair and revokes the old session (rotation). Replaying an already-rotated token is treated as theft and revokes all of the user’s sessions. Token may come from the body or the `refreshToken` cookie.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or reused token."
          }
        },
        "summary": "Rotate the refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Revokes the current session (refresh token).",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "description": "Sets a new password and revokes all other sessions. Accepts either a normal access token or the 5-min `tempToken` from first-login.",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "description": "Creates a single-use 15-min reset token (logged to console in the mock). Always returns success to avoid user enumeration. Limited to 3/hour per user.",
        "operationId": "AuthController_forgot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request a password reset link",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "description": "Consumes the single-use token and sets the new password.",
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password with a token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/businesses/{id}": {
      "get": {
        "description": "Business detail including its active licenses.",
        "operationId": "BusinessController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Business uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The business with active licenses."
          },
          "404": {
            "description": "Business not found."
          }
        },
        "summary": "Get a business by id",
        "tags": [
          "Businesses"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-License Verification Platform API",
    "description": "REST API for the Thai e-licensing platform that plugs into the ทางรัฐ\n(Tang Rat) super app. Serves the public app and the admin portal.\n\n**Authentication.** Most endpoints require a Bearer access token\n(RS256 JWT, 15-min TTL). Obtain one from `POST /api/auth/register` or\n`POST /api/auth/login` (PUBLIC), `POST /api/auth/tang-rat`\n(PUBLIC/INSPECTOR/SUPERVISOR), or `POST /api/auth/self` (ADMIN + TOTP).\nClick **Authorize** and paste the `accessToken` to call protected routes.\nSee `docs/AUTHENTICATION.md` for a plain-English walkthrough.\n\n**Scope.** INSPECTOR/SUPERVISOR responses are filtered server-side by the\ncaller's zones and agency; clients cannot widen scope via query params.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Paste the access token returned by an /auth login route."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "AuthTokenResponseDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "TangRatLoginDto": {
        "type": "object",
        "properties": {}
      },
      "SelfLoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshDto": {
        "type": "object",
        "properties": {}
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}